refactor components and css
