+++ new-dti-website/src/app/apply/page.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import { useState } from 'react';
 import ApplicationTimeline from '../../../components/apply/ApplicationTimeline';
 import config from '../../../config.json';
 import RoleDescriptions from '../../../components/apply/RoleDescription';
 import RedBlob from '../../../components/blob';
+import ApplyFAQ from '../../../components/apply/ApplyFAQ';
 
 const ApplyHero = () => (
   <div className="text-[#FEFEFE] min-h-[calc(100vh-136px)] flex items-center">
@@ -45,173 +45,6 @@ const ApplyHero = () => (
   </div>
 );
 
-const ApplyFAQ = () => {
-  const sections = ['General Questions', 'Behavioral Prep', 'Technical Prep'];
-  const [question, setQuestion] = useState(sections[0]);
-
-  const buttons = sections.map((section) => (
-    <button
-      className={`md:rounded-[30px] xs:rounded-[15px] font-bold md:text-[20px] xs:text-[9px] md:py-4 md:px-5 xs:py-[10px] 
-        xs:px-2 border-[3px] border-[#FEFEFE] ${
-          section === question ? 'text-[#0C0404] bg-[#FEFEFE]' : ''
-        }`}
-      onClick={() => setQuestion(section)}
-    >
-      {section}
-    </button>
-  ));
-
-  const sectionHeadingStyle =
-    'font-semibold md:text-[24px] md:leading-[29px] xs:text-[14px] xs:leading-[17px]';
-  const sectionSubheadingStyle =
-    'font-semibold md:text-[20px] md:leading-[24px] xs:text-[12px] xs:leading-[14px]';
-  const sectionTextStyle = 'md:text-[22px] md:leading-[26px] xs:text-[12px] xs:leading-[14px]';
-  const subsectionStyle = 'flex flex-col gap-2';
-
-  return (
-    <div className="relative flex justify-center text-[#FEFEFE]">
-      <RedBlob className="top-[-300px] left-[-350px] z-0" intensity={0.7} />
-      <div
-        className="flex flex-col max-w-5xl w-full lg:gap-11 md:gap-7 xs:gap-4 
-        lg:px-0 md:px-[60px] xs:px-6 relative z-10"
-      >
-        <h2 className="font-semibold md:text-[32px] xs:text-[22px]">What's next?</h2>
-        <div className="flex flex-col gap-5">
-          <h3 className={sectionHeadingStyle}>
-            Learn more about DTI's core values and processes below.
-          </h3>
-          <div className="flex gap-4">{buttons}</div>
-        </div>
-        <RedBlob className="top-[300px] right-[-500px] z-0" intensity={0.7} />
-        <div className="flex flex-col gap-6">
-          <h3 className={sectionHeadingStyle}>{question}</h3>
-          {question === sections[0] && (
-            <div className="flex flex-col gap-6">
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>Application Review</h3>
-                <p className={sectionTextStyle}>
-                  We look through every single application we receive. We fully understand and don't
-                  expect that everyone has a significant amount of past experience. Taking into
-                  account your year and past coursework, we will assess your interest in being a
-                  part of Cornell Digital Tech & Innovation and your willingness to learn and make a
-                  significant contribution.
-                </p>
-                <p className={sectionTextStyle}>
-                  We have one behavioral and one technical round for each role interview. During
-                  these interviews, talk about your past experience and interest in our team. Then,
-                  we will initiate some role-specific exercises to further gauge your thought
-                  process and skill set.
-                </p>
-              </div>
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>Interviews to Expect</h3>
-                <p className={sectionTextStyle}>
-                  Shortly after applications close, we will offer interviews to candidates we are
-                  interested in. These interviews serve as a chance to get to know you in person. We
-                  have one behavioral and one technical round for each role interview. During these
-                  interviews, talk about your past experience and interest in our team. Then, we
-                  will initiate some role-specific exercises to further gauge your thought process
-                  and skill set.
-                </p>
-              </div>
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>Coffee Chatting</h3>
-                <p className={sectionTextStyle}>
-                  Coffee chats are casual conversations allow you to find out more about DTI and ask
-                  any questions about our team. A “coffee chat” doesn't need to actually be over
-                  coffee, but should be 30 minutes like an actual coffee catch up with a friend. Get
-                  the most out of the coffee chat by preparing your questions ahead of time and
-                  researching the other person's experiences first. Find DTI members to chat{' '}
-                  <a className="underline" href={config.coffeeChatLink}>
-                    here
-                  </a>
-                  .
-                </p>
-              </div>
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>Decisions?</h3>
-                <p className={sectionTextStyle}>
-                  Whether or not you receive an interview invitation, we will email you a definitive
-                  decision within a week of applying! We're happy to answer any questions you have
-                  during this time through our email,{' '}
-                  <a className="underline" href="mailto:hello@cornelldti.org">
-                    hello@cornelldti.org
-                  </a>
-                  .
-                </p>
-              </div>
-            </div>
-          )}
-          {question === sections[1] && (
-            <div className={subsectionStyle}>
-              <h3 className={sectionSubheadingStyle}>What to Expect</h3>
-              <ul className="flex flex-col gap-2 marker:text-[#FEFEFE] list-disc pl-8 space-y-3 ">
-                <li className={sectionTextStyle}>
-                  In our behavioral interviews, you will have the opportunity to share your
-                  experiences and passions with us in a casual setting (no need to dress up!). We
-                  want to get to know what makes you who you are and what brings you to DTI!
-                </li>
-                <li className={sectionTextStyle}>
-                  If selected, you will get a zoom invitation in your email. We, as a team, just
-                  want to get a better idea of you as a candidate, how you work in a team, and what
-                  Cornell DTI means to you.
-                </li>
-                <li className={sectionTextStyle}>
-                  You can find some behavioral practice online or chat with members about what to
-                  expect. This isn't something to stress out about. Believe us, we've all been
-                  there. Above all, we want to get to know you.
-                </li>
-              </ul>
-            </div>
-          )}
-          {question === sections[2] && (
-            <div className="flex flex-col gap-6">
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>Designer</h3>
-                <p className={sectionTextStyle}>
-                  The design technical interview consists of 15 minutes behavioral questions that
-                  are more design focused. Then, a 15 minutes portfolio review + 5 minutes of
-                  questions. You will present a slide deck on one of your past projects. We'll end
-                  with a 25 minute whiteboard challenge where we'll present you with a problem space
-                  and you will create a solution solving it.
-                </p>
-              </div>
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>Developer</h3>
-                <p className={sectionTextStyle}>
-                  To prepare for this interview, we recommend reviewing data structures (OOP) and
-                  system design. You're expected to perform white-board challenges (yes, no code!)
-                  to design features of a website. Ultimately, we want to see how you think and
-                  solve problems.
-                </p>
-              </div>
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>Business</h3>
-                <p className={sectionTextStyle}>
-                  Our technical interview questions will comprise of business case studies
-                  simulating challenges you may face in your role. We do not expect any specific
-                  background knowledge heading into your interview; however, the best way to prepare
-                  would be to thoroughly research the expectations of your role. We look for
-                  candidates that show passion and the ability to make rational decisions with
-                  limited information.
-                </p>
-              </div>
-              <div className={subsectionStyle}>
-                <h3 className={sectionSubheadingStyle}>PM</h3>
-                <p className={sectionTextStyle}>
-                  Our technical interview questions are similar to questions real product managers
-                  get asked in industry. If you google “PM interview questions” you'll find a lot of
-                  helpful resources! We recommend following the CIRCLES framework. We'll also send
-                  materials to help you prepare at that stage.
-                </p>
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
 
 const ApplyCoffeeChat = () => (
   <div className="relative flex justify-center text-[#FEFEFE] md:mb-[200px] xs:mb-[139px]">
