+++ new-dti-website/src/app/globals.css
@@ -74,3 +74,56 @@
     @apply bg-background text-foreground;
   }
 }
+
+.section-heading {
+  font-weight: 600;
+}
+
+/* Mobile phone viewport */
+@media screen and (min-width: 320px) {
+  .section-subheading {
+    font-size: 12px;
+    line-height: 14px;
+  }
+
+  .section-heading {
+    font-size: 14px;
+    line-height: 17px;
+  }
+
+  .section-text {
+    font-size: 12px;
+    line-height: 14px;
+  }
+}
+
+/* Tablet viewport */
+@media screen and (min-width: 768px) {
+  .section-subheading {
+    font-size: 20px;
+    line-height: 24px;
+  }
+
+  .section-heading {
+    font-size: 24px;
+    line-height: 29px;
+  }
+
+  .section-text {
+    font-size: 22px;
+    line-height: 26px;
+  }
+}
+
+/* Laptop viewport */
+@media screen and (min-width: 1024px) {
+
+}
+
+.section-subheading {
+  font-weight: 600;
+}
+
+.section-heading {
+  font-weight: 600;
+}
